Safari-611.1.10/LayoutTests/fast/dom/assign-to-prototype-accessor-on-prototype-should-throw-expected.txt

Makes sure that assigning to a DOM property directly on the prototype throws a TypeError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.prototype.oncopy = "fail" threw exception TypeError: undefined is not an object (evaluating 'document.prototype.oncopy = "fail"').
PASS HTMLElement.prototype.innerText = "fail" threw exception TypeError: The HTMLElement.innerText setter can only be used on instances of HTMLElement.
PASS successfullyParsed is true

TEST COMPLETE
